\name{vegaSpectrum}
\docType{data}
\alias{vegaSpectrum}
\title{Vega CALSPEC spectrum}
\description{
  Vega CALSPEC spectrum alpha_lyr_stis_006

}
\usage{vegaSpectrum}
\format{A data frame with two columns. The first column contains the wavelength in Angstrons. The second column contains the flux in .}
\source{\href{http://www.stsci.edu/hst/observatory/cdbs/calspec.html}{Space Telescope Science Institute 2014}}
\references{

}
\keyword{datasets}



